# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Memoh API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Validate user credentials and issue a JWT",
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/bots/{id}/channel/{platform}": {
            "get": {
                "description": "Get bot channel configuration",
                "tags": [
                    "bots"
                ],
                "summary": "Get bot channel config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/channel.ChannelConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update bot channel configuration",
                "tags": [
                    "bots"
                ],
                "summary": "Update bot channel config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bot ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Channel platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Channel config payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/channel.UpsertConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/channel.ChannelConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/me/channels/{platform}": {
            "get": {
                "description": "Get channel binding configuration for current user",
                "tags": [
                    "channel"
                ],
                "summary": "Get channel user config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/channel.ChannelUserBinding"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update channel binding configuration for current user",
                "tags": [
                    "channel"
                ],
                "summary": "Update channel user config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Channel platform",
                        "name": "platform",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Channel user config payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/channel.UpsertUserConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/channel.ChannelUserBinding"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "channel.ChannelConfig": {
            "type": "object",
            "properties": {
                "botID": {
                    "type": "string"
                },
                "capabilities": {
                    "type": "object",
                    "additionalProperties": true
                },
                "channelType": {
                    "$ref": "#/definitions/channel.ChannelType"
                },
                "createdAt": {
                    "type": "string"
                },
                "credentials": {
                    "type": "object",
                    "additionalProperties": true
                },
                "externalIdentity": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "routing": {
                    "type": "object",
                    "additionalProperties": true
                },
                "selfIdentity": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "verifiedAt": {
                    "type": "string"
                }
            }
        },
        "channel.ChannelType": {
            "type": "string",
            "enum": [
                "telegram",
                "feishu",
                "cli",
                "web"
            ],
            "x-enum-varnames": [
                "ChannelTelegram",
                "ChannelFeishu",
                "ChannelCLI",
                "ChannelWeb"
            ]
        },
        "channel.ChannelUserBinding": {
            "type": "object",
            "properties": {
                "channelType": {
                    "$ref": "#/definitions/channel.ChannelType"
                },
                "config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "channel.UpsertConfigRequest": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "type": "object",
                    "additionalProperties": true
                },
                "credentials": {
                    "type": "object",
                    "additionalProperties": true
                },
                "external_identity": {
                    "type": "string"
                },
                "routing": {
                    "type": "object",
                    "additionalProperties": true
                },
                "self_identity": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "string"
                },
                "verified_at": {
                    "type": "string"
                }
            }
        },
        "channel.UpsertUserConfigRequest": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.LoginRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "handlers.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}